Move e2e config functionality to a separate file
